Cleaned up many warnings, fixed pmap.

+++ core/schedule.c
@@ -222,7 +222,7 @@ struct thread *thread_create(void)
 
 static void __print_fault_info(struct thread *t, int fault, void *info)
 {
-	printk("unhandled fault: %ld: %d\n", t ? t->id : -1, fault);
+	printk("unhandled fault: %ld: %d\n", t ? (long)t->id : -1, fault);
 	debug_print_backtrace();
 	switch(fault) {
 		struct fault_object_info *foi;
